fix(build): only import component types in electron app
