+++ apps/electron-app/src/render/components/react-flow/nodes/Mqtt.tsx
@@ -1,4 +1,8 @@
-import { MqttData, MqttDirection, MqttValueType } from '@microflow/components';
+import type {
+	MqttData,
+	MqttDirection,
+	MqttValueType,
+} from '@microflow/components';
 import { useMqtt } from '@microflow/mqtt/client';
 import {
 	Badge,
