+++ apps/electron-app/src/render/components/react-flow/nodes/matrix/Matrix.tsx
@@ -1,4 +1,4 @@
-import {
+import type {
 	MatrixData,
 	MatrixShape,
 	MatrixValueType,
