+++ apps/electron-app/src/render/components/react-flow/nodes/piezo/Piezo.tsx
@@ -1,4 +1,4 @@
-import {
+import type {
 	BuzzData,
 	PiezoData,
 	PiezoValueType,
