+++ packages/components/src/components/IfElse.ts
@@ -1,22 +1,5 @@
+import { SubValidator, Validator } from '../constants/IfElse';
 import { BaseComponent, BaseComponentOptions } from './BaseComponent';
-
-export const IF_ELSE_VALIDATORS = ['boolean', 'number', 'text'] as const;
-export type Validator = (typeof IF_ELSE_VALIDATORS)[number];
-
-export const IF_ELSE_SUB_VALIDATORS = {
-	boolean: [null],
-	number: [
-		'equal to',
-		'greater than',
-		'less than',
-		'between',
-		'outside',
-		'is even',
-		'is odd',
-	],
-	text: ['equal to', 'includes', 'starts with', 'ends with'], // TODO regex
-} as const;
-export type SubValidator = (typeof IF_ELSE_SUB_VALIDATORS)[Validator][number];
 
 export type IfElseData = {
 	validatorArgs: unknown[];
@@ -30,6 +13,7 @@ export type IfElseOptions = BaseComponentOptions<boolean> & {
 	subValidator?: SubValidator;
 	validatorArgs: any[];
 };
+export type { SubValidator, Validator } from '../constants/IfElse';
 
 export class IfElse extends BaseComponent<IfElseValueType> {
 	constructor(private readonly options: IfElseOptions) {
